UPD RegisterFererred

+++ Functions/Referred/ReferredHistoryRequest.cs
@@ -4,7 +4,7 @@ namespace HeroServer
 {
     public class ReferredHistoryRequest
     {
-        public int AppUserId { get; set; }
+        public long AppUserId { get; set; }
         public DateTime DateStart { get; set; }
         public DateTime DateEnd { get; set; }
 
@@ -12,7 +12,7 @@ namespace HeroServer
         {
         }
 
-        public ReferredHistoryRequest(int appUserId, DateTime dateStart, DateTime dateEnd)
+        public ReferredHistoryRequest(long appUserId, DateTime dateStart, DateTime dateEnd)
         {
             AppUserId = appUserId;
             DateStart = dateStart;
